requirements.txt
debugpy
python-dotenv